requirements.txt
asgiref==3.2.7
astroid==2.4.1
beautifulsoup4==4.9.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.0.6
django-bootstrap4==1.1.1
django-heroku==0.3.1
django-registration==3.1
django-tinymce==3.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.5.2
python-decouple==3.3
pytz==2020.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.0.1
wrapt==1.12.1
